deploy: Supabase PostgreSQL + Vercel config
- vercel.json: clean config for Vercel PHP runtime
- VERCEL.md: Supabase + Vercel deployment guide
- Migration: tinyInteger → smallInteger for PostgreSQL compat
- pdo_pgsql extension confirmed available

+++ database/migrations/2026_09_15_000001_create_comments_and_ratings_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
             $table->id();
             $table->string('author_name');
             $table->string('author_email')->nullable();
-            $table->tinyInteger('stars')->unsigned()->default(5); // 1-5
+            $table->smallInteger('stars')->unsigned()->default(5); // 1-5
             $table->text('review')->nullable();
             $table->nullableMorphs('rateable'); // post ou service
             $table->timestamps();
